Added REST API endpoints to the mechanism for managing custom stop words
